+++ src/Components/Checkout/Cart/CartItems.js
@@ -19,7 +19,6 @@ class CartItems extends Component{
                 .UpdateCartMinus(ID, quantity)
       
                 .then(cart => {
-                  console.log('cart quantity updated');
                   dispatch({ type: FETCH_CART_END, payload: cart, gotNew: true });
                 })
       
@@ -37,7 +36,6 @@ class CartItems extends Component{
                 .UpdateCartPlus(ID, quantity)
       
                 .then(cart => {
-                  console.log('cart quantity updated');
       
                   dispatch({ type: FETCH_CART_END, payload: cart, gotNew: true });
                 })
@@ -56,7 +54,6 @@ class CartItems extends Component{
                 .UpdateCart(ID, quantity)
         
                 .then(cart => {
-                    console.log('cart quantity updated');
                     dispatch({ type: FETCH_CART_END, payload: cart });
                 })
         
@@ -76,7 +73,6 @@ class CartItems extends Component{
           var productArray = products.data.filter(function(product) {
             return product.id === item.product_id;
           });
-
           var product = productArray[0];
 
           var background = product.background_colour;
@@ -128,7 +124,6 @@ class CartItems extends Component{
                       defaultValue={item.quantity}
                       onBlur={event => {
                         cart_edit(item.id, event.target.value);
-                        console.log(event.target.value);
                       }}
                     />
                     <button
