--- src/Components/Checkout/Checkout.scss
@@ -1,1 +1,0 @@
-@import './../variables';
